requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
joblib
shap
streamlit
requests
python-dotenv
